Moving specific value types to subpackage to avoid circular dependencies
