+++ src/mlte/value/types/integer.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 
 from typing import Dict, Any
 
-from .value import Value
+from ..value import Value
 from mlte.validation import Validator, ValidationResult, Success, Failure
 from mlte.measurement_metadata.measurement_metadata import MeasurementMetadata
