+++ test/schema/test_boundspec_schema.py
@@ -9,7 +9,7 @@ Unit tests for BoundSpec schema.
 import mlte
 from mlte.spec import Spec
 from mlte.property.costs import StorageCost
-from mlte.value import Integer
+from mlte.value.types import Integer
 from mlte.measurement.measurement_metadata import MeasurementMetadata
 from mlte.store.api import read_boundspec
 from mlte._private.schema import validate_boundspec_schema
