+++ test/schema/test_result_schema.py
@@ -3,7 +3,7 @@ Unit tests for Value schema.
 """
 
 import mlte
-from mlte.value import Integer, Real, Opaque
+from mlte.value.types import Integer, Real, Opaque
 from mlte.measurement_metadata.measurement_metadata import MeasurementMetadata
 from mlte._private.schema import validate_value_schema
 from mlte.store.api import read_value
